Some corrects

+++ src/main/java/Product.java
@@ -1,12 +1,12 @@
 public class Product {
-    private long id;
+    private Long id;
     private String productName;
-    private float price;
-    private float weight;
+    private Float price;
+    private Float weight;
     private String color;
-    private float productCount;
+    private Integer productCount;
 
-    public Product(long id, String productName, float price, float weight, String color, float productCount) {
+    public Product(Long id, String productName, Float price, Float weight, String color, Integer productCount) {
         this.id = id;
         this.productName = productName;
         this.price = price;
@@ -15,7 +15,7 @@ public class Product {
         this.productCount = productCount;
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
@@ -23,11 +23,11 @@ public class Product {
         return productName;
     }
 
-    public float getPrice() {
+    public Float getPrice() {
         return price;
     }
 
-    public float getWeight() {
+    public Float getWeight() {
         return weight;
     }
 
@@ -35,15 +35,15 @@ public class Product {
         return color;
     }
 
-    public float getProductCount() {
+    public Integer getProductCount() {
         return productCount;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(Float price) {
         this.price = price;
     }
 
-    public void setProductCount(float productCount) {
+    public void setProductCount(Integer productCount) {
         this.productCount = productCount;
     }
 }
